% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about_ui.R
\name{AboutModUI}
\alias{AboutModUI}
\title{About the App Module}
\usage{
AboutModUI(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
UI code to display information about the application
}
\description{
A shiny module that explains the dashboard and certain rules about components that can or cannot be added inside
another element.
}
