% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_code_srv.R, R/mod_code_ui.R
\name{mod_code_server}
\alias{mod_code_server}
\alias{mod_code_ui}
\title{Template Code Module}
\usage{
mod_code_server(id, ui_code)

mod_code_ui(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}

\item{ui_code}{Reactive object containing JSON string of the UI in the "App UI" tab}
}
\value{
UI and server code to display selected HTML elements
}
\description{
Module showing the user the R code required to create the UI on the "App UI" tab.
}
