% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_components.R
\name{component}
\alias{component}
\alias{componentTab}
\title{Component Settings Shell}
\usage{
component(id, ...)

componentTab(id)
}
\arguments{
\item{id}{The ID of the component input}

\item{...}{Shiny tags to include inside the component}
}
\value{
A shiny.tag of the component settings
}
\description{
A container for the specified component input
}
\details{
The tab component contains a selection of specific inputs related to adding a new tab, as
the events to create it in the UI are different to the other components
}
