% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{componentTag}
\alias{componentTag}
\alias{componentType}
\alias{componentPlot}
\alias{componentValue}
\alias{componentLabel}
\alias{componentID}
\alias{componentIcon}
\alias{componentColour}
\alias{componentBackground}
\alias{componentText}
\alias{componentTextArea}
\alias{componentChoices}
\alias{componentRange}
\alias{componentInline}
\alias{componentDownload}
\alias{componentWidth}
\alias{componentHeight}
\alias{componentWidthNum}
\alias{componentOffset}
\alias{component}
\title{Bootstrap Component Inputs}
\usage{
componentTag(ns)

componentType(ns)

componentPlot(ns)

componentValue(ns)

componentLabel(ns)

componentID(ns)

componentIcon(ns)

componentColour(ns)

componentBackground(ns)

componentText(ns)

componentTextArea(ns)

componentChoices(ns)

componentRange(ns)

componentInline(ns)

componentDownload(ns)

componentWidth(ns)

componentHeight(ns)

componentWidthNum(ns)

componentOffset(ns)

component(id, ...)
}
\arguments{
\item{ns}{Namespace to include the component}
}
\value{
A \code{shiny.tag.list} of settings specific to the selected component
}
\description{
A way to be able to adjust components so that can more easily visualise how the shiny application will look.
}
\seealso{
\code{\link{component}}
}
