% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{column_settings}
\alias{column_settings}
\alias{header_settings}
\alias{button_settings}
\alias{dropdown_settings}
\alias{input_settings}
\alias{output_settings}
\title{Bootstrap Component Settings}
\usage{
column_settings(id)

header_settings(id)

button_settings(id)

dropdown_settings(id)

input_settings(id)

output_settings(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
A \code{shiny.tag.list} of settings specific to the selected component
}
\description{
A way to be able to adjust components so that can more easily visualise how the shiny application will look.
}
