% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about_ui.R, R/mod_sidebar_ui.R
\name{mod_about_ui}
\alias{mod_about_ui}
\alias{mod_sidebar_ui}
\title{About the App Module}
\usage{
mod_about_ui(id)

mod_sidebar_ui(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
UI code to display information about the application

UI and server code to display options to create
}
\description{
A shiny module that explains the dashboard and certain rules about components that can or cannot be added inside
another element.

A shiny module that creates Bootstrap components to drag into the App UI canvas for the wire-framing
of shiny applications.
}
