% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designer.R
\docType{package}
\name{designer-package}
\alias{designer}
\alias{designer-package}
\title{designer: 'Shiny' UI Prototype Builder}
\description{
To learn more about \code{designer}, have a read of the vignette: \code{vignette("designer")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbaldry/designer}
  \item \url{https://ashbaldry.shinyapps.io/designer/}
  \item Report bugs at \url{https://github.com/ashbaldry/designer/issues}
}

}
\author{
\strong{Maintainer}: Ashley Baldry \email{arbaldry91@gmail.com}

Other contributors:
\itemize{
  \item Sam Parmar \email{parmartsam@gmail.com} [contributor]
}

}
\keyword{internal}
