% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_pal.R
\name{get_pal}
\alias{get_pal}
\title{generate color palette from Dribble}
\usage{
get_pal(url, pallet = 8)
}
\arguments{
\item{url}{input URL}

\item{pallet}{integer}
}
\description{
generate color palette from Dribble
}

