% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_shot_pal.R
\name{get_shot_pal}
\alias{get_shot_pal}
\title{generate color palette from Dribble shot}
\usage{
get_shot_pal(palette = 8, genre = c("debut", "recent", "teams", "playoffs",
  "animated"), post.n = 1)
}
\description{
generate color palette from Dribble shot
}

