% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_rscript.R
\name{htmlToRScript}
\alias{htmlToRScript}
\title{Convert HTML Info to R Calls}
\usage{
htmlToRScript(html_list, indent = 0)
}
\arguments{
\item{html_list}{A list containing information about the tags, relevant R functions and extra arguments to
give to each HTML tag}

\item{indent}{The number of spaces to indent each subsequent call}
}
\value{
A string that can be written to a \code{ui.R} file
}
\description{
Recursively looking into a list of HTML information to create R code that will produce the HTML seen in the app.
}
