% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_rscript.R
\name{jsonToRScript}
\alias{jsonToRScript}
\title{Convert JSON to R Calls}
\usage{
jsonToRScript(json)
}
\arguments{
\item{json}{A string containing JSON code of the "App UI" page}
}
\value{
A string that can be written to a \code{ui.R} file
}
\description{
Using the JSON string generated from the canvas, convert into a list that can be easily parsed by R
into a script to recreate the UI designed in the app
}
