% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_template_utils.R
\name{save_template}
\alias{save_template}
\title{Store Saved Prototype}
\usage{
save_template(html, page = NULL, title = NULL, desc = NULL, user = NULL)
}
\arguments{
\item{html}{Character string of the HTML that is present in the canvas}

\item{page}{Type of page selected for the template (used to restore correct value on restart)}

\item{title}{Short title of the template}

\item{desc}{Longer description of the template}

\item{user}{Person who created the template}
}
\description{
Saves the current template in the designer cache, along with metadata
}
