% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Component Accordion Item}
\usage{
sidebarItem(id, name, element, parent_id, ..., notes = NULL, active = FALSE)
}
\arguments{
\item{id}{HTML ID to namespace on}

\item{name}{Label to show on the closed accordion}

\item{element}{Character string to let JS know what component has been chosen}

\item{parent_id}{HTML ID of the accordion}

\item{...}{Option inputs to add when expanding the accordion item}

\item{notes}{A list of optional notes to include at the bottom of the settings}

\item{active}{Logical, should the accordion item be open on start? Default set to \code{FALSE}}
}
\value{
A \code{shiny.tag} element containing the component accordion item with all input settings
}
\description{
An item to add to the sidebar that opens up the settings for the selected component
}
