% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{warning_modal}
\alias{warning_modal}
\title{Warning Modal}
\usage{
warning_modal(id)
}
\arguments{
\item{id}{ID to give to the modal}
}
\value{
HTML for a modal
}
\description{
Creates a modal when switching page type in case a user is about to delete the template, giving them a
warning about deleting the page.
}
