% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{columnSettings}
\alias{columnSettings}
\alias{headerSettings}
\alias{buttonSettings}
\alias{dropdownSettings}
\alias{inputSettings}
\alias{outputSettings}
\title{Bootstrap Component Settings}
\usage{
columnSettings(id)

headerSettings(id)

buttonSettings(id)

dropdownSettings(id)

inputSettings(id)

outputSettings(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
A \code{shiny.tag.list} of settings specific to the selected component
}
\description{
A way to be able to adjust components so that can more easily visualise how the shiny application will look.
}
