% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{columnSettings}
\alias{columnSettings}
\alias{rowSettings}
\alias{headerSettings}
\alias{textSettings}
\alias{inputSettings}
\alias{checkboxSettings}
\alias{radioSettings}
\alias{dropdownSettings}
\alias{buttonSettings}
\alias{outputSettings}
\title{Bootstrap Component Settings}
\usage{
columnSettings(id)

rowSettings(id)

headerSettings(id)

textSettings(id)

inputSettings(id)

checkboxSettings(id)

radioSettings(id)

dropdownSettings(id)

buttonSettings(id)

outputSettings(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
A \code{shiny.tag.list} of settings specific to the selected component
}
\description{
A way to be able to adjust components so that can more easily visualise how the shiny application will look.
}
