% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_utils.R
\name{column_settings}
\alias{column_settings}
\alias{header_settings}
\title{Bootstrap Component Settings}
\usage{
column_settings(ns)

header_settings(ns)
}
\arguments{
\item{ns}{Namespace of the overall module}
}
\value{
A \code{shiny.tag.list} of settings specific to the selected component
}
\description{
A way to be able to adjust components so that can more easily visualise how the shiny application will look.
}
