% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_ui.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\title{Bootstrap Component Creation Module}
\usage{
mod_sidebar_ui(id)
}
\arguments{
\item{id}{The character vector to use for the namespace.}
}
\value{
UI and server code to display options to create
}
\description{
A shiny module that creates Bootstrap components to drag into the App UI canvas for the wire-framing
of shiny applications.
}
